added scoreboard class
