+++ lib/game.js
@@ -1,22 +1,24 @@
-var firebase = require('firebase');
-require('firebase/database');
-var config = {
-  apiKey: "AIzaSyB7WYLfPqM1FhRDE6WUEaru-14-d963vN4",
-  authDomain: "go-go-nyan-cat.firebaseapp.com",
-  databaseURL: "https://go-go-nyan-cat.firebaseio.com",
-  storageBucket: "go-go-nyan-cat.appspot.com",
-};
-var firebaseApp = firebase.initializeApp(config);
-var fireDb = firebaseApp.database();
+// var firebase = require('firebase');
+// require('firebase/database');
+// var config = {
+//   apiKey: "AIzaSyB7WYLfPqM1FhRDE6WUEaru-14-d963vN4",
+//   authDomain: "go-go-nyan-cat.firebaseapp.com",
+//   databaseURL: "https://go-go-nyan-cat.firebaseio.com",
+//   storageBucket: "go-go-nyan-cat.appspot.com",
+// };
+// var firebaseApp = firebase.initializeApp(config);
+// var fireDb = firebaseApp.database();
 
 var lifeCounter = 0;
 const Sushi = require('./sushi');
 const Trash = require('./trash');
 const Helpers = require('./helpers');
 const Draw = require('./draw');
+const Scoreboard = require('./scoreboard');
 
 var draw = new Draw();
 var helpers = new Helpers();
+var scoreboard = new Scoreboard();
 var sprites = [];
 var points = 0;
 var endingFrames = 0;
@@ -106,43 +108,42 @@ Game.prototype.endGame = function(gameLoop, context, canvas, hearts) {
   if (endingFrames < 25) {
     requestAnimationFrame(gameLoop);
   } else {
-    this.savePoints();
+    scoreboard.savePoints(points);
+    // this.savePoints();
     this.clearCanvas(context, canvas);
     this.showGameOverScreen();
     this.resetGame(hearts);
   }
 };
 
-Game.prototype.savePoints = function(){
-  var username = $("#username").val() || "unknown user"
-  scoreboardRecords.push({username: username, points: points});
-  fireDb.ref('highscore/').set({
-    highscores: scoreboardRecords
-  });
-};
+// Game.prototype.savePoints = function(){
+//   var username = $("#username").val() || "unknown user"
+//   scoreboardRecords.push({username: username, points: points});
+//   fireDb.ref('highscore/').set({
+//     highscores: scoreboardRecords
+//   });
+// };
 
-var scoreboardRecords = [];
+// var scoreboardRecords = [];
 
-Game.prototype.loadScoreboard = function(){
-  scoreboardRecords = [];
-  $('#scoreboard-records').empty();
-
-  fireDb.ref('highscore/').once('value').then(function(scores){
-    var allScores = helpers.sortScores(scores.val().highscores);
-
-    // var allScores = scores.val().highscores.sort(function(a, b){return b.points - a.points});
-
-    for (var i = 0; i < 5; i++) {
-      if (allScores[i] !== undefined) {
-        scoreboardRecords.push(allScores[i]);
-      }
-    }
-
-    for (var i = 0; i < scoreboardRecords.length; i++) {
-      $('#scoreboard-records').append(i+1 + ". " + allScores[i].username + ": " + allScores[i].points + "<br>");
-    }
-  });
-};
+// Game.prototype.loadScoreboard = function(){
+//   scoreboardRecords = [];
+//   $('#scoreboard-records').empty();
+//
+//   fireDb.ref('highscore/').once('value').then(function(scores){
+//     var allScores = helpers.sortScores(scores.val().highscores);
+//
+//     for (var i = 0; i < 5; i++) {
+//       if (allScores[i] !== undefined) {
+//         scoreboardRecords.push(allScores[i]);
+//       }
+//     }
+//
+//     for (var i = 0; i < scoreboardRecords.length; i++) {
+//       $('#scoreboard-records').append(i+1 + ". " + allScores[i].username + ": " + allScores[i].points + "<br>");
+//     }
+//   });
+// };
 
 Game.prototype.resetGame = function(hearts){
   lifeCounter = 0;
