+++ lib/helpers.js
@@ -56,9 +56,9 @@ Helpers.prototype.checkLoseHeart = function (lifeCounter, hearts) {
   }
 };
 
-Helpers.prototype.sortScores = function (scores){
-  var sortedScores = scores.sort(function(a, b){return b.points - a.points})
-  return sortedScores
-};
+// Helpers.prototype.sortScores = function (scores){
+//   var sortedScores = scores.sort(function(a, b){return b.points - a.points})
+//   return sortedScores
+// };
 
 module.exports = Helpers;
