+++ lib/index.js
@@ -1,5 +1,5 @@
 $(document).ready(function(){
-  game.loadScoreboard();
+  scoreboard.loadScoreboard();
   $("#start-button").on("click", showCanvas);
   $("#restart-button").on("click", showStartScreen);
 });
@@ -13,6 +13,7 @@ const Game = require('./game');
 const Heart = require('./heart');
 const Cat = require('./cat');
 const Background = require('./background');
+const Scoreboard = require('./scoreboard');
 
 var game = new Game();
 var background = new Background({starfield: starfield, starfieldCtx: starfieldCtx, canvas: canvas});
@@ -20,6 +21,7 @@ var nyanCat = new Cat({context: context});
 var heart1 = new Heart(500, {context: context});
 var heart2 = new Heart(550, {context: context});
 var heart3 = new Heart(600, {context: context});
+var scoreboard = new Scoreboard();
 var hearts = [heart1, heart2, heart3];
 var lastGenTime = 0;
 var backgroundImage = background.randomStarsImage(starfieldCtx, canvas, starfield);
@@ -32,7 +34,7 @@ $(document).on('keydown', function(event){
 function showStartScreen(){
   $("#game-over-screen").hide();
   $("#start-screen").show();
-  game.loadScoreboard();
+  scoreboard.loadScoreboard();
 }
 
 function showCanvas(){
